+++ BepInExPlugins/LoadoutLockdown/CHANGELOG.md
@@ -1,4 +1,5 @@
 # 1.1.0
+- Moving an equipped weapon out of a weapon slot will unequip it. (unless its configured to not require a hotbar slot)
 - Added `ApplyPvpMenuSwapRulesToPVE` option. Can be enabled to make PVP menu-swapping rules apply to PVE combat too.
 
 # 1.0.0